fix standard text endpoint
standard text was moved to be decoupled from the text timeseries
