+++ cwms/__init__.py
@@ -10,6 +10,7 @@ from .locations.physical_locations import *
 from .ratings.ratings import *
 from .ratings.ratings_spec import *
 from .ratings.ratings_template import *
+from .standard_text.standard_text import *
 from .timeseries.timeseries import *
 from .timeseries.timeseries_bin import *
 from .timeseries.timeseries_txt import *
